fix cupertino tab scaffold test
