+++ dev/bots/check_tests_cross_imports.dart
@@ -228,7 +228,6 @@ class TestsCrossImportChecker {
   // TODO(justinmc): Fix all of these tests so there are no cross imports.
   // See https://github.com/flutter/flutter/issues/177028.
   static final Set<String> knownCupertinoCrossImports = <String>{
-    'packages/flutter/test/cupertino/material/tab_scaffold_test.dart',
     'packages/flutter/test/cupertino/text_selection_test.dart',
     'packages/flutter/test/cupertino/app_test.dart',
     'packages/flutter/test/cupertino/picker_test.dart',
